Refactor utility module in different files
