+++ src/commands/test.ts
@@ -1,6 +1,6 @@
 import Command from "../model/Command";
 import { CommandScope } from "../model/CommandGroup";
-import { CommandPermission } from "../utils";
+import { CommandPermission } from "../util/discord";
 
 /*
 function test(args: string[], messageRef: Message) {
